refactor:passing props
