+++ src/layout/Layout.jsx
@@ -5,13 +5,26 @@ import Info from '../components/Info/Info'
 import Nav from '../components/Nav/Nav'
 import i18n from '../locales/index.js'
 import JobDetails from '../components/JobDetails/JobDetails'
+import { fetchCatchmentJobs } from "../api/crosscutRequests"
+import { useToggle } from "../hooks/useToggle"
 
 function Layout() {
     const [showCreateModal, setShowCreateModal] = useState(false)
     const [showInfoModal, setShowInfoModal] = useState(false)
     const [showJobDetailsModal, setShowJobDetailsModal] = useState(false)
     const [modalText, setModalText] = useState({ title: "", action: ""})
+    const [jobs, setJobs] = useState()
+    const [isToggled, toggle] = useToggle(false)
 
+
+    useEffect(() => {
+        fetchJobs()
+      }, [isToggled])
+    
+      const fetchJobs = async () => {
+        const resp = await fetchCatchmentJobs()
+        setJobs(resp)
+      }
 
     // handle create modal
     const handleCreate = () => {
@@ -28,11 +41,11 @@ function Layout() {
         setModalText({ title: i18n.t("Catchment details"), action: i18n.t("Close")})
     }
     return <>
-        <Nav handleClick={handleCreate} handleInfo={handleInfo}/>
+        <Nav handleClick={handleCreate} jobs={jobs} handleInfo={handleInfo}/>
         { showCreateModal === true ? <Create title={modalText.title} setShowCreateModal={setShowCreateModal} action={modalText.action}/> : null}
         { showInfoModal === true ? <Info setShowInfoModal={setShowInfoModal}/> : null}
         { showJobDetailsModal === true ? <JobDetails setShowJobDetailsModal={setShowJobDetailsModal} title={modalText.title} action={modalText.action}/> : null}
-        <ListCatchmentJobs handleJobDetails={handleJobDetails}/>
+        <ListCatchmentJobs handleJobDetails={handleJobDetails} jobs={jobs} toggle={toggle}/>
     </>
 }
