+++ src/pages/ListCatchmentJobs.jsx
@@ -9,26 +9,27 @@ import {
 } from "@dhis2/ui";
 import styles from './ListCatchmentJobs.module.css'
 import JobItem from "../components/JobItem/JobItem";
-import { fetchCatchmentJobs } from "../api/crosscutRequests";
-import { useToggle } from "../hooks/useToggle"
+// import { fetchCatchmentJobs } from "../api/crosscutRequests";
+// import { useToggle } from "../hooks/useToggle"
 import i18n from "../locales/index"
 
 function ListCatchmentJobs(props) {
-  const [jobs, setJobs] = useState(null)
-  const [isToggled, toggle] = useToggle(false)
+  const { jobs, toggle, handleJobDetails } = props
+  // const [jobs, setJobs] = useState(null)
+  // const [isToggled, toggle] = useToggle(false)
   const [{ column, direction }, setSortInstructions] = useState({
     column: 'date',
     direction: 'desc',
   })
 
-  useEffect(() => {
-    fetchJobs()
-  }, [isToggled])
+  // useEffect(() => {
+  //   fetchJobs()
+  // }, [isToggled])
 
-  const fetchJobs = async () => {
-    const resp = await fetchCatchmentJobs()
-    setJobs(resp)
-  }
+  // const fetchJobs = async () => {
+  //   const resp = await fetchCatchmentJobs()
+  //   setJobs(resp)
+  // }
 
   // handle sorting of columns
   const getSortDirection = (columnName) => {
@@ -64,7 +65,7 @@ function ListCatchmentJobs(props) {
               if ((direction === 'desc' && strA < strB) ||(direction === 'asc' && strA > strB)) return 1
               return 0
           }).map((job) => {
-            return <JobItem toggle={toggle} key={job.id} name={job.name} status={job.status} id={job.id} date={job.date} handleJobDetails={props.handleJobDetails}/>
+            return <JobItem toggle={toggle} key={job.id} name={job.name} status={job.status} id={job.id} date={job.date} handleJobDetails={handleJobDetails}/>
           })}</TableBody>
         </DataTable>
       </Card>
